Break out fuse-specific things and general FS things
Issue: #2
