--- fs/const.go
@@ -1,13 +1,0 @@
-package fs
-
-const (
-	cardsDir   = "cards"
-	myDecksDir = "my-decks"
-
-	cardImagePrefix  = "image."
-	cardJSONFilename = "card.json"
-
-	deckJSONFilename  = "deck.json"
-	deckCardsDir      = "cards"
-	deckImageFilename = "deck.jpg"
-)
