+++ fsutil/const.go
@@ -1,0 +1,13 @@
+package fsutil
+
+const (
+	CardsDir   = "cards"
+	MyDecksDir = "my-decks"
+
+	CardImagePrefix  = "image."
+	CardJSONFilename = "card.json"
+
+	DeckJSONFilename  = "deck.json"
+	DeckCardsDir      = "cards"
+	DeckImageFilename = "deck.jpg"
+)
