+++ fsutil/image_manager.go
@@ -1,4 +1,4 @@
-package fs
+package fsutil
 
 import (
 	"context"
